renpy: Lock operations that allocate/deallocate SDL_PixelFormats.
This makes it possible for two threads to allocate and deallocate
surfaces with the same format. Previously, if that happened,
SDL_PixelFormat->refcount would not get updated, which could
lead to a crash in SDL_FreeFormat.

Ren'Py can do this in rare circumstances when loading images.
(It was doing it more frequently in video code, but we are
ignoring that now.)